requirements.txt
faiss-cpu
sentence-transformers
transformers
sec-edgar-downloader
beautifulsoup4
nltk
bitsandbytes
accelerate
streamlit
